% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/call_vdat.R
\name{call_vdat}
\alias{call_vdat}
\title{Interface directly with vdat.exe}
\usage{
call_vdat(what = "--help", ...)
}
\arguments{
\item{what}{arguments passed to vdat.exe. Defaults to '--help'}

\item{...}{arguments passed to sys::exec_internal}
}
\description{
Interface directly with vdat.exe
}
