% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{clean_up}
\alias{clean_up}
\alias{rm_docker_image}
\alias{rm_vdat}
\title{Clean up after extracting
Remove ocker images}
\usage{
rm_docker_image()

rm_vdat(x)
}
\arguments{
\item{x}{\code{vdat.exe}}
}
\description{
Clean up after extracting
Remove ocker images

Delete \code{vdat.exe}
}
