% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{error_args}
\alias{error_args}
\alias{error_docker_install}
\alias{error_docker_start}
\alias{error_file_location}
\alias{error_functions}
\alias{error_generic_call}
\alias{error_image_name}
\alias{error_path}
\alias{error_too_many_files}
\title{Error functions}
\usage{
error_args(x, valid_args)

error_docker_install()

error_docker_start()

error_file_location(vdata_file, error)

error_generic_call(what, error)

error_image_name(x)

error_path(x)

error_too_many_files(output_format)
}
\arguments{
\item{x}{argument}

\item{error}{Character. Error from exec_internal(), passed through \code{rawToChar}}

\item{vald_args}{vector of valid arguments}
}
\description{
Error functions
}
\keyword{internal}
