% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{to_csv}
\alias{to_csv}
\title{Convert VRL file(s) to CSV(s)}
\usage{
to_csv(file, outdir = getwd())
}
\arguments{
\item{file}{location of VRL file(s) to convert to CSV.}

\item{outdir}{output directory for the created CSV files. Defaults to the
current working directory.}
}
\description{
Convert VRL file(s) to CSV(s)
}
\examples{
\dontrun{


}

}
