% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{to_folder}
\alias{to_folder}
\title{Convert a VRL file to a folder of CSVs by data type}
\usage{
to_folder(file, outdir = getwd())
}
\arguments{
\item{file}{location of VRL file(s) to convert to CSV.}

\item{outdir}{output directory for the created CSV files. Defaults to the current
working directory.}
}
\description{
Convert a VRL file to a folder of CSVs by data type
}
