% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{utilities}
\alias{utilities}
\alias{skip_on_ci}
\alias{skip_on_runiverse}
\title{Miscellaneous functions for package checking, building, and CI}
\usage{
skip_on_ci()

skip_on_runiverse()
}
\description{
Miscellaneous functions for package checking, building, and CI
}
\section{Details}{


\code{skip_on_ci} and \code{skip_on_runiverse} check the environment for variables
called "CI" and "MY_UNIVERSE", respectively, and return \code{TRUE} if it does
not exist. Used to run examples if the package is being built locally and
there's a chance that \code{vdat.exe} exists. If the package is being built on
a continuous integration platform like GitHub Actions, the "CI" variable
will be \code{TRUE} and \code{skip_on_ci} will return \code{FALSE}. If it is being
built locally, "CI" will be "" and \code{skip_on_ci} will return
\code{TRUE}. Similarly, if the package is being built on R-Universe, the
"MY_UNIVERSE" variable will have your universe's name.
}

