% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vdat_call.R
\name{vdat_call}
\alias{vdat_call}
\title{Interface directly with \code{vdat.exe}.}
\usage{
vdat_call(what = "--help", ...)
}
\arguments{
\item{what}{arguments passed to \code{vdat.exe}. Defaults to \code{--help}.}

\item{...}{arguments passed to \code{sys::\link[sys]{exec_internal}}.}
}
\description{
Low-level interface to \code{vdat.exe}, wherein arguments are passed directly
to \code{vdat.exe}.
}
