% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vdat_format.R
\name{vdat_csv_format}
\alias{vdat_csv_format}
\title{Format object returned by \code{vdat_to_csv}}
\usage{
vdat_csv_format(
  data,
  event_type = NULL,
  event_field = NULL,
  event_desc = FALSE
)
}
\arguments{
\item{data}{dataframe object to be formatted}

\item{event_type}{name of the event type in the detection log to be used in filtering.
Names can be the following \code{"DET"}, \code{"DIAG"}, \code{"HEALTH_VR2W" }, \code{"BATTERY" },
\code{"CLOCK_REF"}, and/or \code{"EVENT_OFFLOAD"}, with the default set to \code{"DET"}.}

\item{event_field}{name of event fields to be assigned as column names.
Default is \code{"DIAG_DESC"}}

\item{event_desc}{if \code{TRUE}, event description dataframe is returned providing
the name of the \code{event_field} corresponding columns associated with that particular
event. Default is \code{FALSE}.}
}
\description{
Format object returned by \code{vdat_to_csv}
}
\examples{
\dontshow{if (all(skip_example_on_ci(), skip_example_on_runiverse())) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}

vdat_csv_format()
\dontshow{\}) # examplesIf}
}
