% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vdat_docker.R
\name{vdat_docker}
\alias{vdat_docker}
\title{Run \code{vdat.exe} in docker}
\usage{
vdat_docker(path, args)
}
\arguments{
\item{path}{file path to \code{vdat.sh} and \code{Fatham_Installer.msi}. These need
to be in the same location. Path has to be a \code{character}.}

\item{args}{either \code{extract} or \code{run} see @details for more information.
\code{args} has to be a \code{character}.}
}
\description{
Allows the ability to pull GitHub Container of vdat-docker using \code{vdat.sh}.
}
\details{
When \code{args = "extract"}, \code{vdat.sh} is provided \code{Fathom_Installer.msi}
which will extract \code{vdat.exe} to project directory.
When \code{args = run},  \code{vdat.sh} is provide \code{vdat.exe} which will
run \code{vdat.exe --help}.
}
