% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vdat_here.R
\name{vdat_here}
\alias{vdat_here}
\title{Write the location of \code{vdat.exe} to the environmental variables.}
\usage{
vdat_here(here = Sys.getenv("RVDAT_EXE"), wine = Sys.getenv("RVDAT_WINE_EXE"))
}
\arguments{
\item{here}{location of \code{vdat.exe}, either a full path name or relative
to the current working directory. Defaults to the value of the "VDAT_EXE"
environment variable}

\item{wine}{location of \code{wine.exe}, either a full path name or relative
to the current working directory. Defaults to the value of the "VDAT_EXE"
environment variable. Only used on Linux or Mac operating systems.}
}
\description{
Write the location of \code{vdat.exe} to the environmental variables.
}
\examples{
\dontshow{if (all(skip_example_on_ci(), skip_example_on_runiverse())) withAutoprint(\{ # examplesIf}

# Auto-find vdat:
vdat_here()

# Provide an explicit location:
vdat_here("c:/program files/innovasea/fathom connect/vdat.exe")
\dontshow{\}) # examplesIf}
}
