% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vdat_inspect.R
\name{vdat_inspect}
\alias{vdat_inspect}
\title{Return general VDAT file metadata}
\usage{
vdat_inspect(vdata_file, ...)
}
\arguments{
\item{vdata_file}{file path of VDAT file to inspect.}

\item{...}{arguments passed to \code{\link{vdat_call}}.}
}
\description{
Return general VDAT file metadata
}
