% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vdat_template.R
\name{vdat_template}
\alias{vdat_template}
\title{Return description fields ("template") of different VDAT data types and/or files}
\usage{
vdat_template(format = "csv.fathom", print = FALSE, ...)
}
\arguments{
\item{format}{file format of the template. Currently only accepts "csv.fathom";
here for future compatibility.}

\item{print}{logical. Defaults to FALS.E Print the output to console?}

\item{...}{arguments passed to \code{\link{vdat_call}}.}
}
\description{
Return description fields ("template") of different VDAT data types and/or files
}
