% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vdat_to.R
\name{vdat_to}
\alias{vdat_to}
\title{Internal \verb{vdat convert} caller}
\usage{
vdat_to(
  output_format = c("csv", "json"),
  vdata_file,
  outdir = getwd(),
  time_corrected = FALSE,
  quiet = FALSE,
  folder = FALSE,
  filter = NULL
)
}
\arguments{
\item{output_format}{What is the desired output format? One of "csv" or "json".
Default is "csv".}

\item{vdata_file}{path of VDAT file to convert to CSV.}

\item{outdir}{output directory for the created CSV files. Defaults to the
current working directory.}

\item{time_corrected}{logical. Do you want to apply the default time correction?}

\item{quiet}{logical, defaults to FALSE. Suppress messages?}

\item{folder}{logical, defaults to FALSE. Convert VDAT to a folder of CSVs?}

\item{filter}{character, defaults to NULL and is currently ignored with a
warning. When implemented, filtering via vdat.exe will only be available for
HR3 receivers.}
}
\description{
Internal \verb{vdat convert} caller
}
