% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vdat_to.R
\name{vdat_to_csv}
\alias{vdat_to_csv}
\title{Convert VRL file(s) to CSV(s)}
\usage{
vdat_to_csv(
  vdata_file,
  outdir = getwd(),
  time_corrected = FALSE,
  quiet = FALSE,
  folder = FALSE
)
}
\arguments{
\item{vdata_file}{path of VRL file to convert to CSV.}

\item{outdir}{output directory for the created CSV files. Defaults to the
current working directory.}

\item{time_corrected}{logical. Do you want to apply the default time correction?}

\item{quiet}{logical, defaults to FALSE. Suppress messages?}

\item{folder}{logical, defaults to FALSE. Convert VDAT to a folder of CSVs?}
}
\description{
Convert VRL file(s) to CSV(s)
}
