% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vdat_to_csv.R
\name{vdat_to_folder}
\alias{vdat_to_folder}
\title{Convert a VDAT file to a folder of CSVs by data type}
\usage{
vdat_to_folder(...)
}
\arguments{
\item{...}{arguments passed to \code{vdat_to}.}
}
\description{
A wrapper around \code{vdat_to_csv} that just changes \code{folder = FALSE} to
\code{folder = TRUE}.
}
\examples{
\dontshow{if (all(skip_example_on_ci(), skip_example_on_runiverse())) withAutoprint(\{ # examplesIf}
\dontshow{\}) # examplesIf}
}
