% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vdat_version.R
\name{vdat_version}
\alias{vdat_version}
\title{Report the version of VDAT identified in \code{\link{vdat_here}}.}
\usage{
vdat_version(...)
}
\arguments{
\item{...}{arguments passed to \code{vdat_call}.}
}
\description{
Report the version of VDAT identified in \code{\link{vdat_here}}.
}
\examples{
\dontshow{if (all(skip_example_on_ci(), skip_example_on_runiverse())) withAutoprint(\{ # examplesIf}

vdat_version()
\dontshow{\}) # examplesIf}
}
