% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/call_vdat.R
\name{call_vdat}
\alias{call_vdat}
\title{Interface directly with \code{vdat.exe}.}
\usage{
call_vdat(what = "--help", ...)
}
\arguments{
\item{what}{arguments passed to \code{vdat.exe}. Defaults to \code{--help}.}

\item{...}{arguments passed to \code{sys::\link[sys]{exec_internal}}.}
}
\description{
Low-level interface to \code{vdat.exe}, wherein arguments are passed directly
to \code{vdat.exe}.
}
