% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{error_functions}
\alias{error_functions}
\alias{error_generic_call}
\alias{error_convert}
\title{Error functions}
\usage{
error_generic_call(what)

error_convert(shell_out, vdata_file)
}
\description{
Error functions
}
\keyword{internal}
