% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{error_functions}
\alias{error_functions}
\alias{error_generic_call}
\alias{error_file_location}
\alias{error_too_many_files}
\title{Error functions}
\usage{
error_generic_call(what, error)

error_file_location(vdata_file, error)

error_too_many_files(output_format)
}
\arguments{
\item{error}{Character. Error from exec_internal(), passed through \code{rawToChar}}
}
\description{
Error functions
}
\keyword{internal}
