% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{interact_docker}
\alias{interact_docker}
\alias{get_image_size}
\alias{image_name}
\alias{image_size}
\alias{start_docker}
\title{Interact with Docker functions}
\usage{
get_image_size(image_name, type)

image_name(image_name = NULL)

image_size(image_name = NULL)

start_docker()
}
\arguments{
\item{image_name}{docker image name}

\item{type}{is either \code{"download_size"} or \code{"unpacked_size"}.}
}
\description{
Interact with Docker functions
}
\keyword{internal}
