% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sys_outputs}
\alias{sys_outputs}
\alias{print.vdat_docker}
\alias{print.vdat_resp}
\title{Methods}
\usage{
\method{print}{vdat_docker}(x, image_size, ...)

\method{print}{vdat_resp}(x, ...)
}
\description{
Methods for VDAT responses
}
\keyword{internal}
