% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vdat_here.R
\name{vdat_here}
\alias{vdat_here}
\title{User-supplied location of vdat.exe}
\usage{
vdat_here(here)
}
\arguments{
\item{here}{location of vdata.exe, either a full path name or relative to the
current working directory}
}
\description{
User-supplied location of vdat.exe
}
