% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vdat_inspect.R
\name{vdat_inspect}
\alias{vdat_inspect}
\title{Return general VDAT file metadata}
\usage{
vdat_inspect(file)
}
\arguments{
\item{file}{file path of VDAT file to inspect.}

\item{...}{arguments passed to \code{\link[sys:exec]{sys::exec_internal}}.}
}
\description{
Return general VDAT file metadata
}
