% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vdat_to.R
\name{vdat_to_csv}
\alias{vdat_to_csv}
\title{Convert VRL file(s) to CSV(s)}
\usage{
vdat_to_csv(file, outdir = getwd(), time_corrected = FALSE)
}
\arguments{
\item{file}{location of VRL file(s) to convert to CSV.}

\item{outdir}{output directory for the created CSV files. Defaults to the
current working directory.}

\item{time_corrected}{logical. Do you want to apply the default time correction?}
}
\description{
Convert VRL file(s) to CSV(s)
}
\examples{
\dontrun{

}

}
