% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vdat_to.R
\name{vdat_to_folder}
\alias{vdat_to_folder}
\title{Convert a VRL file to a folder of CSVs by data type}
\usage{
vdat_to_folder(vdata_file, outdir = getwd(), time_corrected = FALSE)
}
\arguments{
\item{vdata_file}{path of VRL file to convert to CSV.}

\item{outdir}{output directory for the created CSV files. Defaults to the
current working directory.}

\item{time_corrected}{logical. Do you want to apply the default time correction?}
}
\description{
Convert a VRL file to a folder of CSVs by data type
}
\examples{
\dontshow{if (all(skip_on_ci(), skip_on_runiverse())) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
\dontshow{\}) # examplesIf}
}
