% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vdat_to.R
\name{vdat_to_folder}
\alias{vdat_to_folder}
\title{Convert a VRL file to a folder of CSVs by data type}
\usage{
vdat_to_folder(file, outdir = getwd())
}
\arguments{
\item{file}{location of VRL file(s) to convert to CSV.}

\item{outdir}{output directory for the created CSV files. Defaults to the
current working directory.}
}
\description{
Convert a VRL file to a folder of CSVs by data type
}
\examples{
\dontshow{if (not_building_on_ci()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
\dontshow{\}) # examplesIf}
}
