% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vdat_version.R
\name{vdat_version}
\alias{vdat_version}
\title{Report the version of VDAT identified in \code{\link{vdat_here}}.}
\usage{
vdat_version()
}
\description{
Report the version of VDAT identified in \code{\link{vdat_here}}.
}
\examples{
\dontshow{if (any(skip_on_ci(), skip_on_runiverse())) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}

vdat_version()
\dontshow{\}) # examplesIf}
}
